refactor: move undelegate claim proof and circuit to stake crate
