+++ crates/core/component/stake/src/undelegate_claim/mod.rs
@@ -1,5 +1,7 @@
 mod action;
 mod plan;
+mod proof;
 
 pub use action::{UndelegateClaim, UndelegateClaimBody};
 pub use plan::UndelegateClaimPlan;
+pub use proof::{UndelegateClaimCircuit, UndelegateClaimProof};
