+++ crates/core/crypto/src/proofs/groth16.rs
@@ -4,14 +4,12 @@ mod nullifier_derivation;
 mod output;
 mod spend;
 mod traits;
-mod undelegate;
 
 pub use delegator_vote::{DelegatorVoteCircuit, DelegatorVoteProof};
 pub use nullifier_derivation::{NullifierDerivationCircuit, NullifierDerivationProof};
 pub use output::{OutputCircuit, OutputProof};
 pub use spend::{SpendCircuit, SpendProof};
 pub use traits::{ParameterSetup, ProvingKeyExt, VerifyingKeyExt};
-pub use undelegate::{UndelegateClaimCircuit, UndelegateClaimProof};
 
 /// The length of our Groth16 proofs in bytes.
 pub const GROTH16_PROOF_LENGTH_BYTES: usize = 192;
@@ -24,9 +22,7 @@ mod tests {
     use crate::{
         asset,
         keys::{SeedPhrase, SpendKey},
-        rdsa,
-        stake::{IdentityKey, Penalty, UnbondingToken},
-        Address, Amount, Balance, Rseed,
+        Address, Balance, Rseed,
     };
     use ark_ff::UniformRand;
     use ark_groth16::{r1cs_to_qap::LibsnarkReduction, Groth16, ProvingKey, VerifyingKey};
@@ -56,43 +52,6 @@ mod tests {
         any::<[u8; 32]>()
             .prop_map(|bytes| Fr::from_le_bytes_mod_order(&bytes[..]))
             .boxed()
-    }
-
-    proptest! {
-    #![proptest_config(ProptestConfig::with_cases(2))]
-    #[test]
-    fn undelegate_claim_proof_happy_path(validator_randomness in fr_strategy(), balance_blinding in fr_strategy(), value1_amount in 2..200u64, penalty_amount in 0..200u64) {
-            let (pk, vk) = UndelegateClaimCircuit::generate_prepared_test_parameters();
-
-            let sk = rdsa::SigningKey::new_from_field(validator_randomness);
-            let validator_identity = IdentityKey((&sk).into());
-            let unbonding_amount = Amount::from(value1_amount);
-
-            let start_epoch_index = 1;
-            let unbonding_token = UnbondingToken::new(validator_identity, start_epoch_index);
-            let unbonding_id = unbonding_token.id();
-            let penalty = Penalty(penalty_amount);
-            let balance = penalty.balance_for_claim(unbonding_id, unbonding_amount);
-            let balance_commitment = balance.commit(balance_blinding);
-
-            let blinding_r = Fq::rand(&mut OsRng);
-            let blinding_s = Fq::rand(&mut OsRng);
-            let proof = UndelegateClaimProof::prove(
-                blinding_r,
-                blinding_s,
-                &pk,
-                unbonding_amount,
-                balance_blinding,
-                balance_commitment,
-                unbonding_id,
-                penalty
-            )
-            .expect("can create proof");
-
-            let proof_result = proof.verify(&vk, balance_commitment, unbonding_id, penalty);
-
-            assert!(proof_result.is_ok());
-        }
     }
 
     proptest! {
