+++ tools/parameter-setup/Cargo.toml
@@ -7,6 +7,7 @@ publish = false
 [dependencies]
 penumbra-crypto = { path = "../../crates/core/crypto/" }
 penumbra-dex = { path = "../../crates/core/component/dex/" }
+penumbra-stake = { path = "../../crates/core/component/stake/" }
 ark-groth16 = "0.4"
 ark-serialize = "0.4"
 decaf377 = { version = "0.4", features = ["r1cs"] }
